requirements.txt
# Healthcare Prior Authorization POC - skill-driven CrewAI flow
# Use with: .venv\Scripts\activate then pip install -r requirements.txt

crewai>=0.86.0
crewai-tools>=0.17.0
openai>=1.0.0
pypdf>=5.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
requests>=2.28.0
